requirements.txt
Django==1.5.4
MySQL-python==1.2.4
PIL==1.1.7
South==0.8.2
Unipath==1.0
django-braces==1.2.2
django-endless-pagination==2.0
django-taggit==0.10
django-wysiwyg-redactor==0.3.6
pytz==2013d
six==1.4.1
wsgiref==0.1.2
